refactor: extract ImageSlider ui into component
